cleanup BlogPage template and styles
