+++ gatsby-node.ts
@@ -42,7 +42,7 @@ export const onCreateNode = ({ node, actions }) => {
 export const createPages = async ({ graphql, actions }) => {
   const { createPage } = actions;
 
-  const blogTemplate = path.resolve('./src/templates/BlogPage.tsx');
+  const blogTemplate = path.resolve('./src/templates/BlogPage/BlogPage.tsx');
   const projectTemplate = path.resolve('./src/templates/ProjectPage.tsx');
 
   const postOnlyResult = await graphql(`
